assignment3/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.0
yfinance==0.2.36
python-dotenv==0.19.0 